Fix bug w/ publishing ( + test), move copy_from_template tests to their own file
